Implement support for commit comments
Since the webhook payload doesn't provide the commit title, which we use
for the push message subject lines, I had to add an API call to fetch
the title.

This *should* cause gmail to thread commits with comments properly, but
I wasn't able to get sendmail to actually send emails to myself to test
it out.

Still need to split the push email into one email per commit to make
threading perfect.